+++ raphtory/src/algorithms/motifs/mod.rs
@@ -1,4 +1,3 @@
-pub mod three_node_eval_wip;
-pub mod three_node_local;
-pub mod three_node_local_redone;
+pub mod three_node_local_single_thread;
+pub mod three_node_temporal_motifs;
 pub mod three_node_motifs;
